requirements.txt
flask==3.0.0
google-generativeai==0.3.1
instagrapi==2.0.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
huggingface_hub==0.19.4
stability-sdk==0.8.5
openai==1.57.1
gunicorn==21.2.0
